refactor/fix createDiff
